Update the motivation for the new design.
There is a difference between how the engineer working on a project sees things, and how it should be presented to the potential customers.    The engineer sees the problems with the current approach.  The marketing person sees the advantages of the new approach.   Really it is about operating on the thinking of someone who knows almost nothing about the project, as opposed to communicating with someone who knows a lot about the project.  

Hopefully I did not make any technical mistakes (in the line on increased parallelism), and  the engineer in charge of this project likes this marketing presentation.