Merged the original Introduction with the new "Why a new design"
As you suggested. 

it is very common that active developers have less energy for the documentation.  Particularly if it is in a foreign language.  
On the other hand to attract new users, it is critical.  The saying is:  "You only get one chance to make a  good first impression. "

And from a newbies perspective reading the documentation is so much easier than trying to read the source code.  Like a friend of mine said:  "Poor documentation is a show stopper. "

You have done the harder work of creating all of this over the last 5 years.  So I am happy to help some with the documentation.  Thank you for your patience when I get it wrong.